Commit a partial solution to bug #1035. The problem is that once the processes exit they are consideredd "done".  So, the master process is happy to exit as soon as rank 0 exits.  If it happened to have exited due to a signal then the master exits with the rank > 0 procs still running.  Looking at the result of wait() can't tell us if the proc died due to a signal, because the child is the sit() for is ssh, not the app.
The "fix" includes adding a barrier in Finalize (called from gasnet_exit())
so that if ANY process exits before calling Finalize, that can be
distinguished and the master process will due a loop of
"ssh host kill -TERM pid".

The "full" solution, yet to be implemented, will need to check that that the
orphaned processes actually die (rather than just launching the ssh and hoping
for the best).

Also commit a bunch of small mods needed to eventually implement the spawner
such that the master spawns ALL the processes rather than using a tree.  There
is still a little more code needed to complete that work, however.
CVSMAIL: CC: