Implement VAPI scatter/gather support in normal GET/PUT interfaces.
VAPI is capable of scatter/gather only on the initiator.
Specifically, it can scatter TO multiple "segments" on a GET and
gather FROM multiple segments on a PUT.  Each segment is described by
an ordered triple (addr, len, key).  The 'key' is a token returned by
the pinning operation.  A remote addr and key are also supplied as
part of the operation.

Prior to the recent resolution of bug #1028, the SEGMENT_FAST mode of
vapi-conduit prepinned the entire GASNet segment as a single region,
and the pool of bounce buffers as a second pinned region.  Thus every
GET or PUT was using only one of two possible keys, and there was no
use for scatter/gather support for contiguous GETs and PUTs.  However,
the SEGMENT_FAST mode of vapi-conduit now pins the GASNet segment with
several regions of bounded size, and thus may require multiple VAPI
GET/PUT operations for a single GASNet operation.  Where a GASNet
operation spans multiple remote pinned regions, the division into
multiple VAPI operations is unavoidable, because one can only give a
single remote key per operation.  However, when some portion of a
GASNet operation spans a single remote region and multiple local
regions, use of scatter/gather is now used to reduce the operation
count.  This could result in both increased peak bandwidth and
possibly reduced overhead (less overhead in the VAPI library, but more
work done in vapi-conduit).  Actual measurements on pcp-i show the
difference on test{logGP,am,small,large} to be "in the noise".  For
in-segment xfers the fixed size regions mean that an operation
spanning a single remote region can span at most two local regions.
However, for out-of-segment local addresses, the dynamic nature of
firehose means that we can see more than 2 segments.  I've seen 3 in
testing, but 4 is highly unlikely (and may even be impossible because
of the way we work to coalesce adjacent regions).

In SEGMENT_{LARGE,EVERYTHING} we've always used firehose for all of
memory, with no prepinning of the GASNet segment.  We've also tried to
coalesce adjacent regions to minimize the likelihood of splitting a
single GASNet operation into multiple VAPI operations due to spanning
multiple remote regions.  To do so, we had been forcing local pinnings
to align to any hits in the table of remote firehoses.  This meant
that a single page may be covered by many overlapping pinned regions
that were created to align with different remote firehose hits.  Now
we can use scatter/gather to span existing adjacent local pinned
regions without creating redundant pinnings.  Using testvis under
SEGMENT_FAST, I've observed this change in behavior reduced by over
60% the number of calls made to pin memory.  The "volume" of pinned
pages (counting multiple pins of the same page) was reduced by as much
as 75%.  The actual TIME spent in the pinning calls also dropped by
over 60%, but alas the application running time is unchanged.
(Presumably the pinning is fully overlapped with communication, a
supposition supported by the amount of time traces show was spent
stalled on a full send queue.  But that is the result of generating
GET and PUT operations at an unrealistic rate).

Note that testvis was chosen for its irregular memory access patterns
- I don't do any vis-related scatter/gather yet.  Thanks to Dan for
such a brutal tester which easily located some small logic errors I'd
made in the LARGE/EVERYTHING support.  Some in corner cases I'd be
hard pressed to generate test cases for.

As noted above, the largest number of scatter/gather (s/g) segments
observed in my testing (mostly testvis) was 3.  It may even be
impossible to need 4, but I can't prove that right now.  Therefore,
the compiled-in limit is set to 4 segments.  The VAPI interface
requires one to specify the maximum allowable number of s/g segments
at end-point initialization time.  This determines the size of the
data xfered to the HCA each time a VAPI operation is posted.  Evidence
suggests that the data size is rounded up to a power of two, resulting
in a step-function in overhead as the max s/g segment count is
increased.  Fortunately, the first step occurs at 5 (in other words
max_sg=1 and max_sg=4 have equal overhead).  Setting max_sg to 5 or
larger results in an increase in initiator-side-overhead for ALL
operations, whether they use the addition s/g segments or not.  The
max possible is around 30 (It was once 62, but it goes down with each
software release as they use more bits for other things).

When I try to accelerate vis, I figure that 28 s/g segments is going
to be FAR more useful than 4.  To avoid paying an overhead penalty for
non-vis operations, I am thinking I'll create a vis-only end-point
with a high max_sg value and continue using 4 as the limit on the
"normal" end-point.

Correctness has been verified with 'make run-tests-par' on both
pcp-i and SystemX under both FAST and LARGE builds.  EVERYTHING uses
the same code paths as LARGE, but 'make run-tests-par' was done on
a EVERYTHING build on pcp-i as well for good measure.  Jacquard has
been down.

If I'd had the time to build Titanium, I'd have tried the ArrayCopy
test as well.  Dan?