Memory Registration fixes
The error check on register_memory was missing since older versions of the API
had the call return the pointer to the registered memory region and in fact,
the function is still currently documented to do the same.

Add trace information for calls to gm_register_memory.

Until we find a way to cleanly solve this issue, let's make some noise and die
instead of silently allowing a failed registration to continue.

(p.s. commit too late to show in tonight's tests)