Fix bug #1052 (at least for SEGMENT_FAST/SEQ).
As was determined long ago in bug #350 the firehose code violates ANSI
aliasing rules and optimizers must be told of this.

Since I know of at least one PPC/Linux/gm machine at Ames Lab that has xlc,
I have also applied this for gm conduit's CFLAGS when using xlc.

I was able to reproduce the problem with testgasnet on SEQ/FAST.
I've tested this fix only against that case, and found it to be resolved.

If this still affects PAR and/or EVERYTHING please reopen this bug.
Dan can phone me at home this weekend to let me know it has been reopenned.
CVSMAIL: CC: