Add testmpineighbour performance test (MPI analog to testneigbour)
This new mpi-flavoured test compares MPI and UPC in neighbour exchanges.  This
is really a GASNet test since we expect to play with a few different ways to do
the ghost exchange.  Eventually, we'll want to add a strided version of the
neighbour exchange to the GASNet test.

MPI neighbour exchange is inspired from the NAS MG Fortran code, and slightly
tweaked for additional non-blocking performance.

UPC neighbour exchange has different variants.
  1. UPCMG    Current way parry's mg does the ghost exchange
  2. GASNetMG GASNet non-blocking way to do the exchange.  Further optimized
              with waitsome sync flags -- nothing we can't try in UPC but it is
	      fairly complex at the language level.
  3. AMLongMG Active-message based way to do the ghost exchange.  Overall, this
              is the best performer but cannot be expressed in UPC currently.
	      Floating functions should eventually help out.


Also, fix a loop iterations but in testmpiperf.