Fix bug in inttypes detection.
M4 Macro was defining HAVE_STDINT_H, HAVE_INTTYPES_H, etc., to "yes"
instead of to "1".  Now that I've started to use this macro in upcr's
configure, we need this to be correct, else the tests in
upcr_geninclude/stdint.h.in break, since in

    #if HAVE_STDINT_H

HAVE_STDINT_H evaluates to '0' if defined to "yes".

CVSMAIL: CC: