Fix bug 1049 (runtime selection of vapi spawner)
SSH support is always compiled into the executable.
MPI support is compiled in IFF mpi-conduit is built.

At configure time --with-vapi-spawner=mpi|ssh determines which will be used by default.  If this flag is not given then mpi is the default when available, or ssh otherwise.

At runtime the environment variable GASNET_VAPI_SPAWNER will override the configure-time setting.

The option -spawner=mpi|ssh to gasnetrun_vapi takes precedence over the environment and configure-time settings.

If the explicit configure-time setting is MPI when mpi-conduit is disabled, a configure-time error is raised.

If the environment or command line requests MPI as the spawner when mpi-conduit was not compiled, an informative error message is produced.