Fixed bug 1064
The reason for the incompatible VM spaces has been staring me in the face for some time now.

In vapi-conduit, we perform a probe to determine the maximum pinnable memory.  To avoid conflicts among multiple proc on a single node, just one is selected per node to perform the probe and it shares the result with the other nodes.  That is what is causing incompatible VM spaces.

Though there is no mmap() leak, the sequence of mmap() and munmap() calls made to perform the search ends up yielding a different base address than just asking for the same size in a single operation.

Now all nodes go through the mmap() probe, even when they won't be trying to pin it.
CVSMAIL: CC: