Document the reason for the "anomalous" behavior of testqueue on vapi-conduit.
Dan and I noticed the times going up for large puts and gets before reaching the
software queue depth.  Some probing showed that if one changed the queue depth
the anomaly moved too.  Turns out it was just vapi-conduit splitting the large
ops which crossed multiple pinned regions on the remote node and thus consumed
multiple slots in the queue for a single gasnet operation.