Improve resource utilization of firehose w/ vapi-conduit SEGMENT_FAST.
In a SEGMENT_FAST configuration, vapi-conduit uses firehose ONLY for local
pinning, never requesting or granting any remote firehoses.  Yet by default
firehose was reserving 3/4 of the pinning resources for use as remote
firehoses.  To get around this, we explicitly set M and R in the environment
to the smallest values we can get away with, and then tweak the available
resources we report to firehose to recover use of the resources reserved for
the fake M and R.