Commit "Phase I" of work on multi-port/rail vapi support (bug 1095) + This commit resolves all the point-to-point ordering issues raised in   bug 1095, to ensure correct operation when traffic is split over multiple   qps (with no ordering guarantees among them) + This commit does not include any true multi-port of multi-rail use of   vapi hardware.  It simply spreads work over multiple vapi work queues. + All RDMA is load-balanced across available qp's by selecting the one with   the shortest work queue (measured in units of requests, not bytes or any   fancy model).  The actual queue length is out-of-date by the time we use   it but performance (more below) indicates we ARE helping most of the time. + When RDMA must be split into multiple vapi-level operations due to spanning   multiple remote pinned regions, or due to use of bounce buffers, each vapi-   level operation is assigned to a qp independently (a good thing). + AM Requests are load-balanced across available qp's by selecting the one   with the largest number of flow-control credits.  Again the data used will   lag a little in time, but still appears to work. + AM Replies always return via the same qp that the Request arrived on due to   flow-control. + AM Longs w/ non-empty payload always include two or more vapi-level   operations because the payload and header travel separately.  The number   is larger than two when the underlying PUT must be split (into bounce   buffers for non-Async Longs, or remote pinned region alignment for Async   Long).  This is the only GASNet operation in vapi-conduit which still   makes an assuption of point-to-point ordering.  This is achived by doing   the load-balancing choice once (or it is made for us for Replies) and   passing the choice down to the subordinate routines. + In this commit, vapi-conduit is still configured to run a single qp.   Before I will make multi-qp the default, I want to see if anything can be   done to recover the lost throughput of small PUTs (see below).  They are   an important performance point.
Performance summary for 1 vs 2 qps:

Even though we don't make any use of additional hardware (still one port on
one adapter), I am seeing performance improvements when work is striped over
two vapi queue pairs.  This supports the opinion that Dan has held that
enforcing of point-to-point ordering in the network costs something.  The only
real difference in the new code from the way things were done before is to
split the work over work queues that have no ordering guarantees among them.
We actually pay a higher software overhead as a result of dividing the work
this way.  Thus, as far as I can tell, the only reason that performance can
increase is for us to be getting a greater level of concurrency in the network
by providing explicitly independent units of work to the adapter.

For all operations there is an increase in software overhead and/or gap
that is associated with the load-balancing.  Vapi-conduit does some more
work and probably misses cache more often.  The hardware/firmware is
probably also faced with marginally more work when it must manage a larger
number of work queues.

For PUTs through one link-layer MTU (1k) I see a reduction in throughput,
which is consistent with the increased overhead/gap.  The end-to-end
latency does not appear to be affected (within std err).

For PUTs larger than 1k I see throughput increasing.  Keep in mind that with
two qps we can have two operations in flight w/o ordering concerns that
apparently require the final ACK of a PUT before starting the next one in
the same queue.  Even when only one GASNet-level operation is in-flight, it
may become multiple vapi-level operations performed in parallel.  This is
confirmed by reduced end-to-end latency for blocking PUTs that are getting
split into multiple vapi-level operations.

For GETs of ANY size, I see a throughput increase.  Again we can keep more
operations in-flight on the wire by not needing to wait for one operation
to finish before starting the next (where order is now being imposed at the
responder).  For large nb GETs the improvement is significant (>4%).

The story on AM Short and Medium is similar to PUTs.  All Shorts and
Mediums w/ <1k payload see slightly increased latency, but throughput
is seen to improve at all sizes.

For Longs the changes are not very significant.  This partly because a
non-Async Long is cut into 4KB bounce buffers which then get sent over
a single qp to ensure ordering of the payload ahead of the header.  This
is done while blocking until all the bounce buffers are queued.  So, there
is less chance to overlap Longs on multiple qps (memcpy costs more than the
network send, so we have passed more than half the total end-to-end latency
by the time we begin to queue the next Long).  While not tested, I'd guess
that AsyncLong will see improvements like PUT did.