Improve "load-balancing" heuristic for vapi multi-qp support. By splitting the small PUTs over multiple qps, we hurt our performance. However, splitting small GETs helps. The heurustic now places all small PUTs unconditionally on qp0, while large PUTs and all GETs are load-balanced.
Unfortunately, it appears that if the second qp gets used at all, something
is taking place in the vapi drivers or firmware that hurts performance.
For instance, placing all AMs and all PUTs on qp0 yields better testlarge
numbers for small PUTs than if the AMs are moved to qp1.  This makes very
little sense, given that only few AM round trips are used for barriers in
this test.

I have not yet enabled 2-qps as the default because of the following:

Use of 2 qps yields a throughput REDUCUCTION of upto 5% for small PUTs
but a throughput INCREASE of over 30% for small GETs!  I wish there was
a way not to have to compromise like this, but I am seeking input.  If
nobody has good ideas about how to get uniform improvement, I am going to
enable this by default and suffer the small PUT performace loss to gain
the big gain for small GETs (and smaller gains for large PUTs and GETs).