Scaffolding for pthread binding support.
Add check for new linux pthreads library (NTPL).

Allow user to use a non-standard pthreads library (needed to get NPTL on
the Altix and Redhat 9) by specifying PTHREADS_INCLUDE and PTHREADS_LIB.

Parry is going to modify upcr_init to actually bind the pthreads if NPTL
is available AND the user sets an environment setting asking for it (you
don't always want it, esp. for interactive jobs).

CVSMAIL: CC: