Add a configure option to control default ref-extended barrier implementation
Progress functions aren't going to be thoroughly tested in time for the Titanium release, so for now just add a configure option to allow controlling the default extended-ref barrier implementation. Eventually the barrier algorithm decision will become runtime-tweakable with an environment variable (although the default will probably remain configure-selectable).

The options are configure --with-barrier=AMDISSEM or --with-barrier=AMCENTRAL, with AMDISSEM set as the current default because it gives a huge scalability benefit for blocking barriers (eg 7x difference on 64 nodes of Jacquard), at some cost in latency for non-blocking barriers and tiny node counts (15us vs 9us for 2 nodes of Jacquard, cross-over around 4 nodes).

Add an ident string to identify which barrier implementation is in use.
CVSMAIL: CC: