Add initial support for cross-compilation support into GASNet, UPCR and Titanium
After some investigation with Imran, we've determined that BG/L absolutely requires cross-compilation for all compute-node software (because no computer in the system can run both the compute node compiler and all the executables it creates). After poking around a bit I decided this actually wasn't that hard to do, because it only affects configure checks that actually try to execute a built program, which occurs fairly rarely (for exactly the cross-compilation reason).

The new configure instructions are as follows:

* Systems where the host and target are the same platform (ie all the system we could handle up until now) are completely unchanged. Just configure as usual.

* Cross-compilation systems:

  1. Build the program gasnet/other/cross-configure-help.c using the target compiler (the one that builds
     executables for your compute nodes). If compilation fails, try tweaking
     one of the test control variables in that file (and you'll need to manually
     indicate the result for that test). This program basically precomputes all the
     runtime values that configure will need and outputs a script that feeds the
     canned answers to configure.
  2. Run the built program on one of the compute nodes and save the output
     into a file in the top-level source directory named "cross-configure".
  3. Set the new script to be executable: chmod +x cross-configure
  4. Edit the "cross-configure" script for completeness, notably setting the
     full path to your target compilers.
  5. Run cross-configure with the same options you'd pass to configure
     (eg. see cross-configure --help)

UPCR will still need a bit more work to truly support cross-compilation, because some of the C binaries we build using $CC (namely detect-upc) are used in the UPC compilation process, so those presumably will need to be built using a "special" C compiler that can build frontend/host exectuables instead of compute node/target exectuables (requiring the existence of such a compiler on the frontend machine, which may not always be available - I've encountered at least one cross-compilation supercomputer where that was the case - there was no host compiler, in order to prevent confusing users who always should be compiling for the target). Given this is the only dependency, it might warrant rewriting detect-upc using Perl (isn't there a lex for perl?).

Titanium will probably also need a bit of additional work, because tc and all the libraries it uses will need to be built with a host C++ compiler. Luckily, with --disable-amudp the C++ compiler is only used for building tc (and the host/target nodes on BG/L can both handle PPC32 objects), so the hope is that on BG/L we may be able to get away with just setting CXX to the host compiler and CC to the target compiler - if that fails we may need some additional hacking. Finally, the garbage collector is not currently setup to allow cross-compilation, so it should be disabled whenever cross compiling - that's probably OK because most systems that need a cross compiler are probably too exotic to have GC support anyhow.


CVSMAIL: CC: