Fix alignment bug in testnbr that cased SIGBUS on X1
The bad alignment of a single structure caused the X1 to SIGBUS.  This is not
surprising since the benchmark actually carves out its space within the GASNet
segment as opposed to using a real malloc -- GASNet has no malloc.  In this
case, carving out by hand caused a structure's first element, an 8-byte word,
to be aligned on a 4-byte boundary.

Note that this alignment problem is orthogonal to the other SIGBUS failure
observed in bug 1050, for which i have another fix as part of a larger set of
shmem-conduit fixes I'm almost ready to commit.