probable fix to bug 1185
I believe this failure was caused by a change introduced in the Linux 2.6 kernel that alters the format of /proc/meminfo (which incidentally, has apparently broken many pieces of Linux software, in addition to ours).

As a result of this change, shmem-conduit's v2.1.0 gasnetc_getMaxMem() function fails to obtain the memory size of the machine during bootstrap and things just snowball from there.
CVSMAIL: CC: mpm@sgi.com