Apply a better fix for bug 436 - upcc-sizes failures preventing all compilations on the X1 We now have a secondary mechanism for obtaining the upcc sizes information, which is enabled by default with Cray C, and optionally on other platforms using "configure --disable-object-strings". In summary, the new trick works by building a dummy executable per conduit config during libupcr build, scanning the strings out of the executable into a generated per conduit config header file (where they are declared using the X1-friendly GASNETT_IDENT()), and the header is subsequently #included into the libupcr build to generate scannable symbols. Move all the sizes declaration goop unchanged into its own .c file to make this possible.
Note that an independent problem remains for the X1 - both phoenix and robin are apparently lacking a version of 'nm' which is capable of reading the X1 version of libupcr.a, and consequently upcc still fails on these machines. However, sparrow.ccs.ornl.gov (the Solaris compile server, only reachable from ORNL machines) *does* have a working nm and cross-compilations on that system are fast and functional.
CVSMAIL: CC: