apply workaround for bug 361 Paul's diagnosis is correct, and gasnet_mmap.c:769 is the one and only problematic line of code. However his proposed solution cannot be applied at this time because bootstrapExchange is currently not required/implemented for all conduits.
The problematic line of code is only relevant for shmem-conduit (which doesn't support pthreads) so I've applied a temporary workaround to only use that line of code for shmem-conduit - which effectively fixes the problem for now because UPCR will never pass differing seg sizes across nodes of shmem-conduit.

Later (probably after the release) I intend to add bootstrapExchange to the required list in template conduit, update existing conduits appropriately and apply a better fix to this bug.

CVSMAIL: CC: