Final cleanups of (lightly tested) vapi multi-rail support. + Now have env var control over HCAs/ports to be used + Now have configure-time enable/disable for multi-rail (on by default) + Now have configure-time option to increase max HCA count above default of 2 + README now documents the related env vars GASNET_{VAPI_PORTS,NUM_QPS} + Default behavior is now to set num_qps equal to number of active ports
Testing so far has been done by rigging the open code to open the same HCA
twice.  This was sufficient to find one bug yesterday.  It is also expected
to be a pretty good test, as the two instances of the hca are assigned
distinct "protection domains", ensuring the memory keys associated with one
are invalid with the other (assuming a bug-free implementation of the IB
spec by Mellanox).

Next week I'll move cables and cards to test first multiple ports on a
single HCA, followed by multiple HCAs.  I expect no big problems.