More on vapi multi-rail.
This commit improves the code path beween the "bind" and the "post" by
eliminating the possibility of binding to a QP w/ no space.  In particular
the spin for QP space is now done at "bind" time rather than "post" time.
Though these are typically only separated by initialization of a few
struct fields, placing the looping at "bind" means we can "rebind" if
the initial choice turns out to have no space.

At least for now the "most space" heuristic for binding is replaced with a very
simple round-robin.  This change may or may not be permanent.