Fix bug 1389 I am fixing this with a slightly "smaller hammer" than I proposed last night.  It appears sufficient from my experiments to add a compiler fence to the FAST_ALIGNED_MEMCPY macro.  For safety, I am adding 2: one at the start and one at the end.  This should (I think) be sufficient to ensure the compiler doesn't make any reordering of loads or stores across this macro.
I am doing the same to the GASNETE_VALUE_ASSIGN macro because it has the
potential to cause the same problems.

Reopen this bug if you think this fix may be insufficient.
In particular, we could enable the fenced version only for gcc, and use
plain memcpy() for other compilers if we think it necessary.
CVSMAIL: CC: