implement a better fix to bug 1389
prevent bad optimizations on GASNETE_FAST_ALIGNED_MEMCPY and GASNETE_VALUE_ASSIGN due to ansi-aliasing rules added in C99 that foolishly outlaw type-punning.

Exploit a union of all possible base types of the given size as a loophole in the rules to make the assignment legal and prevent "bad" optimizations, with a hopefully minimal impact on "good" optimizations.

 Other possible solutions include forcing gasneti_compiler_fence before&after the type-punning (which would probably hurt optimization on any code surrounding a local access), globally disabling ansi aliasing using compiler flags (bad, because it would globally hurt application code optimization), or redundantly copying the
  first byte of the value as a (char *) to try and fool the optimizer into treating the "illegal" assignment in a safe manner (works for this particular example, but is not guaranteed to work in general, and may introduce redundant and possibly more expensive byte load/store instructions)

CVSMAIL: CC: