fix bug 1358: incompatable VM spaces on mu.hcs.ufl.edu
I've finally had a chance to poke at this bug, apologies for the long delay.

A similar problem is also affecting lem.lbl.gov, a SPARC/Solaris 2.10 machine, where all multi-process jobs fail with the mmap init death. This confirms the problem is not Linux specific or even x86/Opteron specific. This machine is an underpowered little uniprocessor, and provided the missing piece of evidence that was missing for a proper diagnosis.

It turns out the problem is apparently caused because we were unintentionally forgetting to pass MAP_NORESERVE during our mmap search on Linux and Solaris, and consequently relying on the system default behavior, which apparently is to allocate swap space, at least on the systems mentioned in this bug. The restriction for backing swap space causes non-determinism in the mmap search when a given node is low on swap space, eg due to a small configured swap space, possibly in concert with other processes consuming swap memory on a particular node (probably the issue on the UFL machines), or (even worse) due to cross-process interference between GASNet processes running on the same physical node and all performing their swap-backed mmap searches simultaneously (the problem on lem).

The fix is to add MAP_NORESERVE on these OS's (and any others that support the option where we were missing it, notably FreeBSD), and I've confirmed that this appears completely effective in solving the problem on mu (elan, mpi, udp), lambda (vapi, udp) and lem (udp).

Given the severity of the problem and the small size of the fix, I'm very tempted to include this in the upcoming patch release. However, given the wide variety of Linux systems where we run, I'm somewhat nervous to deploy such a fundamental change without some extensive testing, especially for systems with pinning-based cluster network support. Thoughts are appreciated...

When this fix is deployed at UFL, the system default GASNET_MAX_SEGSIZE setting on the UFL systems should become unnecessary, and should probably be removed (to prevent artificially limiting the shared heap space available to Berkeley UPC programs).


CVSMAIL: CC: leko@hcs.ufl.edu