fix bug 1490 - missing ident strings with HP C (head-only fix)
The problem was HP C with optimizations enabled is apparently now smart enough to discard the ident strings embedded in *_startup_tmp.c, but not those in libupcr and libgasnet (which is what you saw). upcrun looks specifically for those in *_startup_tmp.c (which is why it failed).

Fixed by adding an HP C pragma in GASNETT_IDENT to ensure our string embedding works properly. The patch is shown below - the fix is #ifdef'd to only be enabled for HP-C so it should *not* affect other platforms, and therefore *should* be safe to merge in to the release (provided Jason is willing to retest on HP C) - although I'm hesitant to make any source changes at all this late in the testing cycle.

Opinions?

CVSMAIL: CC: