add an AMMPI workaround for another IBM MPI bug
Under heavy ISend/IRecv load (mpi-conduit/testlarge) IBM MPI on seaborg,
datastar, bassi and bassidev were all observed to violate the MPI spec's
ordering semantics and match a send to an irecv which was not the next recv
in the FIFO ordering in which they were posted, causing a deadlock in AMMPI.

This deadlock was also observed using the old AMMPI code (2.2.2 release), so
this is not a new bug, although my recent AMMPI tuning seems to have made it
more likely to be enountered in practice.

On AIX we now detect this deadlock at runtime and automatically drop into a degraded
mode of operation that does not rely upon MPI ordering guarantees (and
consequently pays an increased cost to poll for message arrival which grows
roughly linearly with queue depth).

CVSMAIL: CC: