Add "annotation" of memory barrier properties. This commit does *NOT* change anything about the membar implementations.
These annotations will be used by the fenced atomics code to allow the
preprocessor to eliminated doubled fences on platforms where rmb() and/or
wmb() are a full memory fence.
These will also be used to allow preprocess-time removal of conditionals
that guard no-op memory fences.  This is intended to encourage inlining,
because most (of not all) compilers perform inlining decisions based on
a function's "size" before it can perform the optimizations that would
eventually eliminate the dead code.