initial commit of token-based flow-control for AMMPI
Similar to the vapi-conduit approach, we keep a per-target count of unacknowledged
requests, to bound the number of requests that may be sitting in an unexpected
message queue at the target. Tokens are returned by piggybacking up to 255
returned tokens on both requests and replies, and in one-way request-only
patterns are coalesced at the target up to a threshold where an autoreply is
generated.  Unlike vapi, the recv tokens are not statically partitioned between
nodes, because AMMPI has a shared recv queue.

Various settings can be used to tweak the token counts and coalescing
parameters, which by default maximize token count while guaranteeing no
unexpected messages for strictly pairwise communication patterns and coalesce
up to 75% of the tokens at the target. For N-to-one communication patterns,
the unexpected message queue length is bounded by 2*depth*(N-1), but the depth
and token counts can be tuned to lower this bound when desirable.

This appears to provide uniform performance and stability improvements
over all platforms tested, most notably eliminating the portals crashes
on XT3, at least for pairwise communication patterns.

more improvements to come..
CVSMAIL: CC: