Commit the initial version of atomic add and subtract support, including timings in testmisc and correctness checks in testtools.
There is not yet any platform-specific tuning, as the only implementation
currently available is in terms of compare-and-swap.

There has also been no effort yet to convert the callers to use these,
and there won't be until testmisc and testtools show the (not yet
implemented) platform-specific versions to be "OK".  In this way I can
get the code into nightly testing in a way that only breaks testmisc
and/or testtools if I get the code wrong (short of syntax errors).