Add AMUDP support for cross-platform spawning
Cross-platform spawning is required for systems like RAMP Blue, where the compute node OS is a microkernel and the master console needs to run on a frontend machine of a different architecture.
Add an amudprun program, which can be used to launch any AMUDP or udp-conduit binary.
Rewrite the AMUDP master-slave wire-level protocol to make all messages insensitive to the platform endianness, bit width and structure packing conventions.
Fix a recently-introduced bug in the networkpid generation.
Tweak amx tests to work with new spawn support.

CVSMAIL: CC: