Mostly fix bug 1413.
Disregard what said earlier about having been using "lwsync" all along. We have not been doing so.
However, what I said about "lwsync" executing as "sync" on older processors is correct.  As is the
observation that wmb (as either lwsync or sync) is sufficiently strong for a full mb().

So, this commit uses lwsync for mwb() and mb() on all PPC platforms, while continuing to use isync as
the rmb().

I have verified that the behavior appears correct by using the additions I made to testtools today.
These changes use mb() in the role of either wmb() or rmb() to make sure it has the sufficient
properties.  These tests were conducted on the following: bassi (Power5), seaborg (Power3), k421 (G5),
starsky (G5) and chiptest5 (G4?).

On all five of these platforms the mb() cost is reduced by at least 15% and by as much as 93% (that's
14X faster)!

On all platforms the use of lwsync for wmb() results in either statistically identical timing
(chiptest5 and seaborg) or significant improvements (over 2X on bassi and 11X on k421 and starsky).

On the G5s there is still room for improvement by using lwsync in place of isync for the rmb().  Tests
confirm that to do so is both correct, and reduces the rmb() cost by 2X on the G5 CPU.  However, on
the other CPUs isync is still faster than lwsync.  Because this is still a possible improvement, I am
marking as "REMIND".
CVSMAIL: CC: