AMMPI latency improvement
reduce mpi-conduit RT latency on a number of platforms by moving IRecv buffer re-post
out of the critical put/get latency path and into the wire delay slot immediately following
a send operation, thus improving the likelihood that we overlap the recv post
cost with communication, to get some overlap benefit even in algorithms with no client-level overlap.

This provides a large mpi-conduit RT latency benefit (about 2us) on sandia MPI,
and nearly the same benefit (about 1.7us) on the XT-3. It also provides about a
2us improvement on thunder/elan MPI, and a small but significant benefit (0.5us)
on pcp/MVAPICH.  No noticeable difference on BGL.

The algorithm allows up to (depth-1) recv preposts to be lazily delayed until
the next send, but no benefit was observed for allowing > 1 (and there's reason
to believe it hurts because it temporarily reduces the effective recv queue depth),
so the default slack parameter is 1.

CVSMAIL: CC: