Fix (again) our asm for 64-bit atomics for ILP32 builds on ppc64 cpus. This commit fixes the testtools failures observed on Starsky last night.
It turns out that while Linux (and blrts) use one ABI, Darwin and AIX are
using a different one.  This wouldn't have even been an issue if XLC had
"real" asm support.  However, since XLC asm must know the argument-passing
convention, we need distinct versions for the two ABIs.

"The nice thing about standards is that there are so many to choose from"
  - Andrew S. Tanenbaum

Note that blrts is currently a moot point, since today's BlueGene/L systems
don't have 64-bit cpus.  However, for the future it helps to know that it
is using the Linux ABI (not the AIX one).