Fix missing #include <sys/stat.h> that broke gasnet on most systems.
This caused most of our nightly testers to fail last night.

I moved the include of sys/stat.h from a conditional (if ALTIX) block,
to the top level, so all systems #include it.  This is done
unconditionally, but I assume it's a standard POSIX header, so we don't
need to do any configure.in goop to check for it and define
HAVE_SYS_STAT_H (gasnet seems to do less of these checks than upcr,
anyway).

The file that actually broke from omitting sys/stat.h was
gasnet_tools.c, but it includes gasnet_timers.h, so I figured this was
the easiest way to make the fix.

CVSMAIL: CC: