Distinguish the asm macro used for building "special" function bodies.
This is in preparation for improvements related to bug 1621 and 1752
in which the PGI compiler will end up using both "old" and "new" asm
styles in the same build.  Because the bug that prevents use of GNU-style
asm() for atomics doesn't have any effect on inline asm() for timers, this
will allow inline timers w/ the PGI 6.1-x compilers (via GNU-style asm)
while still implementing "special" atomics via "old" asm() w/o arguments.

Note that this commit changes ALL use of GASNETI_ASM() for "special"
function bodies, but that only PGI 6.2-x will see a change in the
preprocessed source that results.