partial fix to bug 1753
versioning issue in portable_platform is now fixed.

The omp.h header in the 32-bit version apparently changed in the interval <5.2-4,6.0-8], so in 32-bit mode portable_platform.h now reports corrected version numbers (ie <= 5.2-4 or >=6.0-8).

I've verified that all versions of 32/64 bit pgcc and pgCC installed on hive report version numbers which are never incorrect, although depending on the available information only ranges can be provided in some cases.

This is correct, although less useful, because the whole point of the omp.h hack was to detect the first version supporting GNU __asm__ (v 6.1.1). Unfortunately, there does not appear to be a usable #define analogous to the omp.h that would allow us to safely detect the 6.0->6.1 transition in 32-bit mode. However, given the codegen bugs in pgcc inline assembly in versions 6.1.x which prevent us from using it on those versions, I don't think failure to detect 6.1.1 for 32-bit mode actually makes any significant difference for the code we generate.

Version detection is now fixed, and I've verified the HEAD of testtools and testtoolscxx pass on all the 32-bit versions installed on hive in debug and opt modes.

Reassigning bug to Paul to consider rolling back his patch to gasnet_asm.h - portable_platform.h for 32-bit pgi will now never report a version number in the 6.1.* range, so the change no longer has any affect on the 32-bit code it was intended to fix. Furthermore, it might conceivably cause a performance regression for the timers on 64-bit pgcc v6.1.1, whose version is correctly detected, but forced to use special mode by Paul's patch (incurring a function call overhead).
CVSMAIL: CC: